generate ai

+++ Backend/api/handlers/article.py
@@ -11,11 +11,11 @@ from api.schemas import ArticleCreateHistory, UpdateArticleRequest, ShowArticle
 
 from api.utils.task import _save_task_and_articles_images
 from api.utils.jwt import _get_current_user_from_access_token
-from api.utils.article import _create_new_article, _create_history_article, _delete_article, _get_all_article, _update_article, _get_all_history_article
+from api.utils.article import _create_new_article, _create_history_article, _delete_article, _get_all_article, _update_article, _get_all_history_article, _get_article_by_id
 from api.utils.user import _get_user_by_login
 
 from db.session import get_db
-from db.models import User
+from db.models import User, PortalRole
 
 
 article_router = APIRouter()
@@ -65,23 +65,26 @@ async def delete_article(
     id: UUID,
     user: User = Depends(_get_current_user_from_access_token),
     session: AsyncSession = Depends(get_db)) -> Union[None, UUID]:
-   delete_article = await _delete_article(id=id, session=session)
-   if delete_article is None:
-      raise HTTPException(status_code=404, detail=f"Article with id {id} is not found.")
-   article_history = await _create_history_article(
-    ArticleCreateHistory(
-    article_id=delete_article.id_article,
-    article_title=delete_article.title,
-    user_login=user.login,
-    change_event=["Delete article."],
-    timestamp=delete_article.updated_at,
-    ),
-    session=session,
-   )   
-   if article_history is None:
-      raise HTTPException(status_code=503, detail="Failed to create task history.")
-   
-   return delete_article.id
+   article = await _get_article_by_id(id=id, session=session)
+   if article.author_id == user.id or user.role == PortalRole.ROLE_PORTAL_ADMIN:
+      delete_article = await _delete_article(id=id, session=session)
+      if delete_article is None:
+         raise HTTPException(status_code=404, detail=f"Article with id {id} is not found.")
+      article_history = await _create_history_article(
+      ArticleCreateHistory(
+      article_id=delete_article.id_article,
+      article_title=delete_article.title,
+      user_login=user.login,
+      change_event=["Delete article."],
+      timestamp=delete_article.updated_at,
+      ),
+      session=session,
+      )   
+      if article_history is None:
+         raise HTTPException(status_code=503, detail="Failed to create task history.")
+      return delete_article.id
+   else:
+      raise HTTPException(status_code=403, detail="Forbidden.")  
 
 @article_router.patch('/', response_model=Union[UUID, None])
 async def update_article(
@@ -92,25 +95,29 @@ async def update_article(
    update_article_params = body.dict(exclude_none=True)
    if update_article_params == {}:
       raise HTTPException(status_code=422, detail="At least one parameter for article update info should be provided")
-   update_article = await _update_article(id=id, update_article_params=update_article_params, session=session)
-   if update_article is None:
-      raise HTTPException(status_code=404, detail=f"Article with id '{id}' is not found")
-   events_list = []
-   for event in update_article_params:
-      events_list.append(f'Update {event}')
-   article_history = await _create_history_article(
-    ArticleCreateHistory(
-    article_id=update_article.id_article,
-    article_title=update_article.title,
-    user_login=user.login,
-    change_event=events_list,
-    timestamp=update_article.updated_at,
-    ),
-    session=session,
-   ) 
-   if article_history is None:
-     raise HTTPException(status_code=503, detail="Failed to create Article history.")
-   return update_article.id
+   article = await _get_article_by_id(id=id, session=session)
+   if article.author_id == user.id or user.role == PortalRole.ROLE_PORTAL_ADMIN:
+      update_article = await _update_article(id=id, update_article_params=update_article_params, session=session)
+      if update_article is None:
+         raise HTTPException(status_code=404, detail=f"Article with id '{id}' is not found")
+      events_list = []
+      for event in update_article_params:
+         events_list.append(f'Update {event}')
+      article_history = await _create_history_article(
+      ArticleCreateHistory(
+      article_id=update_article.id_article,
+      article_title=update_article.title,
+      user_login=user.login,
+      change_event=events_list,
+      timestamp=update_article.updated_at,
+      ),
+      session=session,
+      ) 
+      if article_history is None:
+         raise HTTPException(status_code=503, detail="Failed to create Article history.")
+      return update_article.id
+   else:
+      raise HTTPException(status_code=403, detail="Forbidden.") 
 
 @article_router.get('/')
 async def get_all_article(user: User = Depends(_get_current_user_from_access_token), session: AsyncSession = Depends(get_db)):
